Move route modules out of subfolders
